ComponentMap and ValidatorMap get passed in
