--- src/validators/validateTruthy.js
@@ -1,3 +1,0 @@
-export default function validateTruthy(x) {
-  return !!x ? '' : 'Please provide input'
-}
